+++ packages/bundle/src/speech/createMicrophoneAudioConfigAndAudioContext.ts
@@ -52,7 +52,10 @@ function createMicrophoneAudioConfig(options: MicrophoneAudioInputStreamOptions)
   const { audioConstraints, audioContext, debug, enableTelemetry, pcmRecorderWorkletUrl } = options;
   const bufferDurationInMS = options.bufferDurationInMS || DEFAULT_BUFFER_DURATION_IN_MS;
 
-  const pcmRecorder = new PcmRecorder(true);
+  // Related to #4523.
+  // When bumping to recent version of `microsoft-cognitiveservices-speech-sdk@>=1.23.0`, pass `true` to the constructor.
+  // const pcmRecorder = new PcmRecorder(true);
+  const pcmRecorder = new PcmRecorder();
 
   pcmRecorderWorkletUrl && pcmRecorder.setWorkletUrl(pcmRecorderWorkletUrl);